requirements.txt
numpy
pandas
Pillow
scikit_learn==1.1.1
scrcpy-client
ipykernel
tqdm
opencv-python
ipywidgets
python-telegram-bot
pytesseract